ADAS_WHEELHUB MOTORS _OS
DONE NOW IT CAN  CONTROL THE WHM WITH SPI AFTER TAKING INPUT FROM BLUETOOTH MODULE AND SENDING IT TO MASTR SPI NODE

+++ ATMEGA32_projects/OS_SCHEDULER_Version1/OS/OS_private.h
@@ -10,7 +10,7 @@
 #define OS_PRIVATE_H_
 
 typedef struct{
-	u8 Periodicity;
+	u16 Periodicity;
 	u8 TaskID;
 	void (*Fptr)(void);
 }Task;
